+++ src/Extensions/Numbers.php
@@ -93,6 +93,6 @@ class Numbers implements IComparer
             return $r;
         }
 
-        return StrictObject::compare($left, $right);
+        return Objects::compare($left, $right);
     }
 }