+++ src/Extensions/Text.php
@@ -255,6 +255,6 @@ class Text extends TextBase implements IComparer
             return $r;
         }
 
-        return StrictObject::compare($left, $right);
+        return Objects::compare($left, $right);
     }
 }