+++ tests/TestCase/Extensions/ObjectsTest.php
@@ -2,37 +2,33 @@
 /**
  * PHP: Nelson Martell Library file
  *
- * Copyright © 2016-2019 Nelson Martell (http://nelson6e65.github.io)
+ * Copyright © 2019 Nelson Martell (http://nelson6e65.github.io)
  *
  * Licensed under The MIT License (MIT)
  * For full copyright and license information, please see the LICENSE
  * Redistributions of files must retain the above copyright notice.
  *
- * @copyright 2016-2019 Nelson Martell
+ * @copyright 2019 Nelson Martell
  * @link      http://nelson6e65.github.io/php_nml/
- * @since     0.6.0
+ * @since     1.0.0
  * @license   http://www.opensource.org/licenses/mit-license.php The MIT License (MIT)
  * */
 
-namespace NelsonMartell\Test\DataProviders;
+namespace NelsonMartell\Test\TestCase\Extensions;
 
-use NelsonMartell\StrictObject;
+use NelsonMartell\Test\DataProviders\Extensions\ObjectsTestProvider;
 
-use NelsonMartell\Test\Helpers\ExporterPlugin;
-use NelsonMartell\Test\Helpers\IComparerTester;
+use PHPUnit\Framework\TestCase;
 
 /**
  *
  * @author Nelson Martell <nelson6e65@gmail.com>
- * @since  0.6.0
  * @internal
+ * @group Criticals
+ *
+ * @since 1.0.0
  * */
-trait StrictObjectTestProvider
+class ObjectsTest extends TestCase
 {
-    use ExporterPlugin;
-
-    public function getTargetClassName() : string
-    {
-        return StrictObject::class;
-    }
+    use ObjectsTestProvider;
 }